+++ _common/gameModel/libs/ClientScript/game/common/actions/actionTemplate/actionTemplateBase.ts
@@ -355,9 +355,3 @@ export class FullyConfigurableChoiceActionTemplate<
     );
   }
 }
-
-// -------------------------------------------------------------------------------------------------
-// -------------------------------------------------------------------------------------------------
-//
-// -------------------------------------------------------------------------------------------------
-// -------------------------------------------------------------------------------------------------
